+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Master version (11/11/2016)
+## Master version (16/11/2016)
 
 - Added the option to to *not* store the full value of pdfs in the output of optimization/evaluation, by setting
 ```python
@@ -8,7 +8,13 @@ glob.annz["doStorePdfBins"] = False
 ```
 In this case, only the average metrics of a pdf are included in the output.
 
+- Added the `sampleFrac_errKNN` option, to allow to sub-sample the input dataset for the knn uncertainty calculation (similar to e.g., `sampleFracInp_wgtKNN` and `sampleFracInp_inTrain`).
+
+- Added metric plots of the distribution of the KNN error estimator vs. the true bias. The plots are added to the output by setting `glob.annz["doKnnErrPlots"] = True`.
+
 - Updated `README.md`.
+
+- Other minor modifications and bug fixes.
 
 ## ANNZ v2.2.1 (01/11/2016)